requirements.txt
openai>=1.0.0
python-dotenv>=0.19.0
anthropic>=0.3.0
huggingface-hub>=0.17.0
numpy>=1.21.0
pandas>=1.3.0
tqdm>=4.62.0
